03.01.2024

+++ src/main/java/ru/app/draft/models/Order.java
@@ -4,7 +4,6 @@ import lombok.Builder;
 import lombok.Data;
 
 import java.math.BigDecimal;
-import java.util.Date;
 
 @Data
 @Builder
@@ -13,14 +12,14 @@ public class Order {
 
     public Order() {}
 
-    public Order(BigDecimal price, Long quantity, String direction, String date) {
+    public Order(BigDecimal price, BigDecimal quantity, String direction, String date) {
         this.price = price;
         this.quantity = quantity;
         this.direction = direction;
         this.date = date;
     }
 
-    private Long quantity;
+    private BigDecimal quantity;
     private String direction;
     private String date;
 }
